* Podcast support.
git-svn-id: https://aqualung.svn.sourceforge.net/svnroot/aqualung/trunk@795 3cd24cdc-1f22-0410-b8b1-dcf80e670293

+++ src/utils_gui.h
@@ -76,7 +76,7 @@ void insert_label_spin(GtkWidget * table, char * ltext, GtkWidget ** spin,
 		       int spinval, int y1, int y2);
 
 void insert_label_spin_with_limits(GtkWidget * table, char * ltext, GtkWidget ** spin,
-				   int spinval, int min, int max, int y1, int y2);
+				   double spinval, double min, double max, int y1, int y2);
 
 void set_option_from_toggle(GtkWidget * widget, int * opt);
 void set_option_from_combo(GtkWidget * widget, int * opt);
